platform: tegra: mc: update sid type to int
Update sid type to int to fix the coverity issue.

CID 63165683
Bug 3461002

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2775857
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>

+++ drivers/platform/tegra/mc/tegra-mc-sid.c
@@ -44,7 +44,7 @@ static LIST_HEAD(sid_override_list);
 struct tegra_mc_sid_override {
 	struct list_head list;
 	void __iomem *addr;
-	unsigned long sid;
+	int sid;
 };
 
 struct tegra_mc_sid {
